source/server/requirements.txt
Django==4.1.3
django-cors-headers==4.0.0
pymysql==1.1.0
cryptography==41.0.7
openai==0.28.1
requests==2.31.0
PyJWT==2.8.0
django-redis==5.4.0
redis==5.0.1
django-ratelimit==4.1.0
pandas==2.0.3
openpyxl==3.1.2
gunicorn==21.2.0
python-dotenv==1.0.0
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
uritemplate==4.2.0
prometheus-client==0.19.0
pytest==7.4.3
pytest-django==4.5.2
pytest-cov==4.1.0
flake8==6.1.0
Pillow==10.1.0
